refactor: 원시값(distance) 랩핑, 일급 컬렉션(Sections) 적용 (피드백 반영)

+++ src/main/java/nextstep/subway/line/dto/LineRequest.java
@@ -13,9 +13,9 @@ public class LineRequest {
     private String color;
     private Long upStationId;
     private Long downStationId;
-    private int distance;
+    private Long distance;
 
-    public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
+    public LineRequest(String name, String color, Long upStationId, Long downStationId, Long distance) {
         this.name = name;
         this.color = color;
         this.upStationId = upStationId;
